Add --type opt to differentiate Hardware and VM's
Previously the script combined both Hardware and VirtualGuests lists
into a single one.  But subsequent API calls for the two server types
are slightly different.

The -T option, set to "Hardware" by default for backward compatibility,
will fetch the desired list, and call the correct API target based on
that type.  This also works with --list.